+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * failure to download and setup a dataset will no longer result in leftover
     dataset files that keep pysaliency from trying again.
   * crossvalidation splits now support stratifying stimulus attributes
+  * the MIT1003 dataset now also contains the history of fixation durations
 * 0.2.19:
   * added pytorch implementation for optimization of similarity metric as alternative
     to tensorflow implementation which still uses tensorflow 1.x